create value.rs
